+++ map.py
@@ -13,6 +13,7 @@ global map  # globalize map to be used in other files
 class MapTile:
     """Map Class to create tiles for the game"""
     def __init__(self, name, desc, event, x, y):
+        self.searchFlag = False
         self.name = name
         self.desc = desc
         self.event = event
@@ -22,8 +23,8 @@ class MapTile:
     def __str__(self):
         return self.name
 
-    def search(self):
-        return self.event
+    def modifyPlayer(self, player):
+        pass
 
     def adjacentMoves(self, maxX, maxY):
         moves = []
@@ -47,6 +48,10 @@ class MapTile:
         moves = self.defaultActions(maxX, maxY)
         return moves
 
+    def search(self):
+        self.searchFlag = True
+        print(self.event)
+
     def getName(self):
         return self.name
 
@@ -62,39 +67,38 @@ class WinTile(MapTile):
     def __init__(self):
         self.searchFlag = False
         self.player = None
+        self.open = False
         super().__init__(name = "locked door",
-                         desc = "fill in desc",
+                         desc = "You see a locked door with three locks",
                          event = """
-                         You see a locked door with three locks, do you wish to open the door?
+                         Upon examination it looks like each lock needs a key to open the door,
+                         do you wish to attempt open the door?
                          """,
                          x = 3,
                          y = 2)
 
-    def search(self):
-        self.searchFlag = True
-        return self.event
-
-    def openDoor(self, player):
-        self.player = player
-        if self.searchFlag == True:
+    def modifyPlayer(self, player):
+        if self.open:
             if player.inventory.count("key") == 3:
                 player.win = True
-                return """
+                print("""
                 Congratulation on finding three keys
                 You win!
-                """
+                """)
             else:
-                return """
+                print("""
                 You need to collect 3 keys to unlock this door
-                """
-        else:
-            return None
+                """)
+
+    def openDoor(self):
+        self.open = True
+
 
     def availableActions(self, maxX, maxY):
         moves = self.defaultActions(maxX, maxY)
         moves.append(act.WinTileSearch())
         if self.searchFlag == True:
-            moves.append(act.WinTileOpenDoor(self.player))
+            moves.append(act.WinTileOpenDoor())
         return moves
 
 winTile = WinTile()
@@ -125,8 +129,9 @@ class EndTile(MapTile):
     """Class to create the "giving up" tile"""
     def __init__(self):
         self.searchFlag = False
+        self.open = False
         super().__init__(name = "exit",
-                         desc = "fill in desc",
+                         desc = "You see a bright exit sign on top of a door",
                          event = """
                          You see a door with an exit sign on top, do you wish to open the door?
                          WARNING - THIS WILL BE TREATED AS YOU GIVING UP ON THE GAME - WARNING
@@ -134,25 +139,33 @@ class EndTile(MapTile):
                          x = 3,
                          y = 0)
 
-    def search(self):
-        self.searchFlag = True
-        return self.event
+    def modifyPlayer(self, player):
+        if self.open:
+            player.giveUp = True
+            print("""
+            You lose
+                """)
 
-    def openDoor(self, player):
+    def openDoor(self):
+        self.open = True
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.EndTileSearch())
         if self.searchFlag == True:
-            self.giveUp = False
-            return """
-            Thank you for playing, you lose
-            """
+            moves.append(act.EndTileOpenDoor())
+        return moves
 
 endTile = EndTile()
 
 class CabinetTile(MapTile):
     def __init__(self, item):
-        self.searchFlag = False
         self.item = item
+        self.openDrawerFlag = False
+        self.searchFlag = False
+        self.keyFlag = False
         super().__init__(name = "cabinet",
-                         desc = "fill in desc",
+                         desc = "You see a brown cabinet with two doors open",
                          event = """
                          You see a generic cabinet with some clothes hanged,
                          a mirror and a drawer. What would you like to do?
@@ -160,43 +173,51 @@ class CabinetTile(MapTile):
                          x = 0,
                          y = 0)
 
-    def search(self):
-        self.searchFlag = True
-        return self.event
+    def modifyPlayer(self, player):
+        if self.openDrawerFlag:
+            if self.keyFlag:
+                print("""
+                you opened the drawer and see nothing
+                """)
+            else:
+                print("""
+                You opened the drawer and found a key!
+                """)
+                player.inventory.append("key")
+                self.keyFlag = True
+                self.openDrawerFlag = False
 
     def searchClothes(self):
-        if self.searchFlag == True:
-            return """
-            You searched through the clothes and noticed some Balenciaga,
-            Supreme, Gucci, Louis Voutton, but didn't find any keys at all.
-            """
-        else:
-            return None
+        print("""
+        You searched through the clothes and noticed some Balenciaga,
+        Supreme, Gucci, Louis Voutton, but didn't find any keys at all.
+        """)
 
     def lookMirror(self):
-        if self.searchFlag == True:
-            return """
-            You look in the mirror and notice a reflection that resembles yourself
-            """
-        else:
-            return None
+        print("""
+        You look in the mirror and notice a reflection that resembles yourself
+        """)
 
-    def openDrawer(self, player):
+    def openDrawer(self):
+        self.openDrawerFlag = True
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.CabinetTileSearch())
         if self.searchFlag == True:
-            player.inventory.append(self.item)
-            return """
-            You open the drawer and found a key!
-            """
-        else:
-            return None
+            moves.extend([act.CabinetTileSearchClothes(), act.CabinetTileLookMirror(), act.CabinetTileOpenDrawer()])
+        return moves
+
 
 cabinetTile = CabinetTile("key")
 
 class ChestTile(MapTile):
     def __init__(self, item):
         self.item = item
+        self.openChestFlag = False
+        self.openFlag = False
         super().__init__(name = "chest",
-                         desc = "fill in desc",
+                         desc = "you see a chest in the corner, covered with dust",
                          event = """
                          You see a chest
                          What would you like to do?
@@ -204,26 +225,45 @@ class ChestTile(MapTile):
                          x = 1,
                          y = 0)
 
-    def openChest(self, player):
-        if "chest key" in player.inventory:
-            player.inventory.append(self.item)
-            player.inventory.remove("chest key")
-            return """
-            You opened the chest and found a key
-            """
-        else:
-            return """
-            You need a key to open this chest.
-            """
+    def modifyPlayer(self, player):
+        if self.openChestFlag:
+            if self.openFlag:
+                print("""
+                You opened the chest and see nothing
+                """)
+            else:
+                if "chest key" in player.inventory:
+                    print("""
+                    You opened the chest and found a key!
+                    """)
+                    player.inventory.append("key")
+                    player.inventory.remove("chest key")
+                    self.openChestFlag = False
+                    self.openFlag = True
+                else:
+                    print("""
+                    You need to find the chest key to open this chest
+                    """)
+
+    def openChest(self):
+        self.openChestFlag = True
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.ChestTileSearch())
+        if self.searchFlag == True:
+            moves.append(act.ChestTileOpenChest())
+        return moves
 
 chestTile = ChestTile("key")
 
 class BookcaseTile(MapTile):
     def __init__(self, item):
         self.item = item
-        self.searchFlag = False
+        self.openBookFlag = False
+        self.keyFlag = False
         super().__init__(name = "bookcase",
-                         desc = "fill in desc",
+                         desc = "you see a bookcase standing tall and strong",
                          event = """
                          You see a bookcase filled with a variety of books with the
                          most common color being red
@@ -231,176 +271,218 @@ class BookcaseTile(MapTile):
                          x = 2,
                          y = 0)
 
-    def search(self):
-        self.searchFlag = True
-        return self.event
+    def modifyPlayer(self, player):
+        if self.openBookFlag:
+            if self.keyFlag:
+                None
+            else:
+                print("""
+                As you opened the book a key fell out!
+                """)
+                player.inventory.append("key")
+                self.openBookFlag = False
+                self.keyFlag = True
 
     def readBook1(self):
-        if self.searchFlag == True:
-            return """
-            Minecraft: Combat Handbook: Ultimate Collector's Edition by Erik Aronsen
-                The Masterpiece from Amazon #1 Bestselling Minecraft Authors Creative Community.
-                This time we're delighted to present A Stunning Master Work - Minecraft Combat
-                Handbook:Collector's Edition Our goal is to show you the most Incredible
-                Possibilities and unlockyour Creative Abilities to Master Minecraft World
-                together with us!
-            """
-        else:
-            return None
+        print("""
+        Minecraft: Combat Handbook: Ultimate Collector's Edition by Erik Aronsen
+            The Masterpiece from Amazon #1 Bestselling Minecraft Authors Creative Community.
+            This time we're delighted to present A Stunning Master Work - Minecraft Combat
+            Handbook:Collector's Edition Our goal is to show you the most Incredible
+            Possibilities and unlockyour Creative Abilities to Master Minecraft World
+            together with us!
+            """)
 
     def readBook2(self):
-        if self.searchFlag == True:
-            return """
-            THE ESSENTIAL CALVIN AND HOBBES by Bill Watterson
-                Beginning with the day Hobbes sprang into Calvin's tuna fish trap, the first
-                two Calvin and Hobbes collections, Calvin and Hobbes and Something Under The Bed
-                Is Drooling, are brought together in this treasury. Including black-and-white
-                dailies and color Sundays, The Essential Calvin and Hobbes also features an
-                original full-color 16-page story.
-            """
-        else:
-            return None
+        print("""
+        THE ESSENTIAL CALVIN AND HOBBES by Bill Watterson
+            Beginning with the day Hobbes sprang into Calvin's tuna fish trap, the first
+            two Calvin and Hobbes collections, Calvin and Hobbes and Something Under The Bed
+            Is Drooling, are brought together in this treasury. Including black-and-white
+            dailies and color Sundays, The Essential Calvin and Hobbes also features an
+            original full-color 16-page story.
+            """)
 
     def readBook3(self):
-        if self.searchFlag == True:
-            return """
-            Diary of a Wimpy Kid Book 1 by Jeff Kinney
-                It’s a new school year, and Greg Heffley finds himself thrust into middle school,
-                where undersized weaklings share the hallways with kids who are taller, meaner,
-                and already shaving. The hazards of growing up before you’re ready are uniquely
-                revealed through words and drawings as Greg records them in his diary.
-            """
-        else:
-            return None
+        print("""
+        Diary of a Wimpy Kid Book 1 by Jeff Kinney
+            It’s a new school year, and Greg Heffley finds himself thrust into middle school,
+            where undersized weaklings share the hallways with kids who are taller, meaner,
+            and already shaving. The hazards of growing up before you’re ready are uniquely
+            revealed through words and drawings as Greg records them in his diary.
+            """)
 
     def readBook4(self):
-        if self.searchFlag == True:
-            return """
+            print("""
             Goosebumps: Welcome to Camp Nightmare by R L Stine
                 The food isn't great. The counselors are strange. And the camp director seems
                 demented. Okay, so Billy can handle all that. But then his fellow campers start
                 to disappear. What's going on? Why won't his parents answer his letters? What's
                 lurking out there after dark? Camp Nightmoon is turning into Camp Nightmare! And
                 Billy might be next...
-            """
-        else:
-            return None
+                """)
 
-    def readBook5(self, player):
-        if self.searchFlag == True:
-            player.inventory.append(self.item[0])
-            return """
-            Little Red Riding Hood by Gaby Goldsack
-                he classic tale of Little Red Riding Hood comes to life in this vibrant retelling
-                perfect for beginning readers.  Designed to encourage vocabulary development and
-                help children read aloud, this story uses larger font types and vivid, contemporary
-                illustrations to help early learning skills. It's a perfect addition to any children's
-                library.
+    def readBook5(self):
+        self.openBookFlag = True
+        print("""
+        Little Red Riding Hood by Gaby Goldsack
+            he classic tale of Little Red Riding Hood comes to life in this vibrant retelling
+            perfect for beginning readers.  Designed to encourage vocabulary development and
+            help children read aloud, this story uses larger font types and vivid, contemporary
+            illustrations to help early learning skills. It's a perfect addition to any children's
+            library
+            """)
 
-                As you opened the book a key fell out!
-            """
-        else:
-            return None
 
     def readBook6(self):
+        print("""
+        Captain Underpants and the Tyrannical Retaliation of the Turbo Toilet 2000 by Dav Pilkey
+            Just when you thought it was safe to flush . . .
+
+            The Turbo Toilet 2000 strikes back! The carnivorous commode known for devouring everything
+            in its path has built up a real appetite . . . for REVENGE! Join Captain Underpants for another
+            epic showdown of Wedgie Power vs. Potty Power as our tighty-whitey-wearing superhero GOES TO ELEVEN!
+            """)
+
+    def readBook7(self):
+        print("""
+        The 39 Clues Book One: The Maze of Bones by Rick Riordan
+            Minutes before she died Grace Cahill changed her will, leaving her decendants an impossible decision:
+            "You have a choice - one million dollars or a clue."
+            """)
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.BookcaseTileSearch())
         if self.searchFlag == True:
-            return """
-            Captain Underpants and the Tyrannical Retaliation of the Turbo Toilet 2000 by Dav Pilkey
-                Just when you thought it was safe to flush . . .
+            moves.extend([act.BookcaseTileRead1(), act.BookcaseTileRead2(), act.BookcaseTileRead3(), act.BookcaseTileRead4(), act.BookcaseTileRead5(), act.BookcaseTileRead6(), act.BookcaseTileRead7()])
+        return moves
 
-                The Turbo Toilet 2000 strikes back! The carnivorous commode known for devouring everything
-                in its path has built up a real appetite . . . for REVENGE! Join Captain Underpants for another
-                epic showdown of Wedgie Power vs. Potty Power as our tighty-whitey-wearing superhero GOES TO ELEVEN!
-            """
-        else:
-            return None
-
-    def readBook7(self, player):
-        if self.searchFlag == True:
-            player.inventory.append(self.item[1])
-            return """
-            The 39 Clues Book One: The Maze of Bones by Rick Riordan
-                Minutes before she died Grace Cahill changed her will, leaving her decendants an impossible decision:
-                "You have a choice - one million dollars or a clue."
-
-                You found a 39 clue card!
-            """
-        else:
-            return None
-
-bookcaseTile = BookcaseTile(["key", "Card # 6: Thomas Jefferson"])
+bookcaseTile = BookcaseTile("key")
 
 class PaintingTile(MapTile):
     def __init__(self):
         super().__init__(name = "painting",
-                         desc = "fill in desc",
+                         desc = "You see a classical painting of doge",
                          event = """
-                         You see a painting
+                         it's just a painting, did you expect something cool to happen?
                          """,
                          x = 0,
                          y = 1)
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.PaintingTileSearch())
+        return moves
 
 paintingTile = PaintingTile()
 
 class DeskTile(MapTile):
     def __init__(self):
         super().__init__(name = "desk",
-                        desc = "fill in desc",
+                        desc = "You see a desk ",
                         event = """
-                        You see a desk
+                        The desk has papers all over it, must be an AP student
                         """,
                         x = 2,
                         y = 1)
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.DeskTileSearch())
+        return moves
 
 deskTile = DeskTile()
 
 class EmptyTile(MapTile):
     def __init__(self):
         super().__init__(name = "",
-                         desc = "fill in desc",
+                         desc = "",
                          event = """
                          There is nothing to see here
                          """,
                          x = 3,
                          y = 1)
 
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.EmptyTileSearch())
+        return moves
+
 emptyTile = EmptyTile()
 
 class ShelveTile(MapTile):
     def __init__(self):
         super().__init__(name = "shelve",
-                         desc = "fill in desc",
+                         desc = "You see a shelve floating on the wall",
                          event = """
-                         You see a shelve
+                         Looks like a strong shelf
                          """,
                          x = 0,
                          y = 2)
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.ShelveTileSearch())
+        return moves
 
 shelveTile = ShelveTile()
 
 class TableTile(MapTile):
     def __init__(self):
         super().__init__(name = "table",
-                        desc = "fill in desc",
+                        desc = "You see a table",
                         event = """
-                        You see a table
+                        The table appears to have four legs
                         """,
                         x = 1,
                         y = 2)
 
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.TableTileSearch())
+        return moves
+
 tableTile = TableTile()
 
 class ChairTile(MapTile):
-    def __init__(self):
+    def __init__(self, item):
+        self.item = item
+        self.sitFlag = False
+        self.keyFlag = False
         super().__init__(name = "chair",
-                         desc = "fill in desc",
+                         desc = "You see a chair sitting in the corner",
                          event = """
-                         You see a chair
+                         It seems very sturdy
                          """,
                          x = 2,
                          y= 2)
 
-chairTile = ChairTile()
+    def modifyPlayer(self, player):
+        if self.sitFlag:
+            if self.keyFlag:
+                print("""
+                The chair made squeeked a little when you sat on it
+                """)
+            else:
+                player.inventory.append(self.item)
+                print("""
+                Upon sitting on the chair you notice that it wobbles a little
+                and found a key for a chest underneath the legs
+                """)
+                self.keyFlag = True
+
+    def sit(self):
+        self.sitFlag = True
+
+    def availableActions(self, maxX, maxY):
+        moves = self.defaultActions(maxX, maxY)
+        moves.append(act.ChairTileSearch())
+        if self.searchFlag:
+            moves.append(act.ChairTileSit())
+        return moves
+
+chairTile = ChairTile("chest key")
 
 map =[
     [cabinetTile, chestTile, bookcaseTile, endTile],